# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_HashanEranga_go-task-manager-service_internal_models.AuthResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/github_com_HashanEranga_go-task-manager-service_internal_models.User'
    type: object
...
  github_com_HashanEranga_go-task-manager-service_internal_models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  github_com_HashanEranga_go-task-manager-service_internal_models.Permission:
    properties:
      action:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      resource:
        type: string
    type: object
  github_com_HashanEranga_go-task-manager-service_internal_models.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  github_com_HashanEranga_go-task-manager-service_internal_models.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        maxLength: 100
        type: string
      last_name:
        maxLength: 100
        type: string
      password:
        minLength: 8
        type: string
      phone:
        maxLength: 20
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  github_com_HashanEranga_go-task-manager-service_internal_models.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/github_com_HashanEranga_go-task-manager-service_internal_models.Permission'
        type: array
      updated_at:
        type: string
    type: object
...
  github_com_HashanEranga_go-task-manager-service_internal_models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified_at:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_email_verified:
        type: boolean
      last_login_at:
        type: string
      last_name:
        type: string
      locked_until:
        type: string
      password_changed_at:
        type: string
      phone:
        type: string
      roles:
        items:
          $ref: '#/definitions/github_com_HashanEranga_go-task-manager-service_internal_models.Role'
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  github_com_HashanEranga_go-task-manager-service_internal_models.UserResponse:
    properties:
      permissions:
        items:
          type: string
        type: array
      roles:
        items:
          type: string
        type: array
      user:
        $ref: '#/definitions/github_com_HashanEranga_go-task-manager-service_internal_models.User'
    type: object
host: localhost:8080
info:
  contact:
    email: support@taskflow.com
    name: API Support
  description: Task management system with JWT authentication and RBAC authorization
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: TaskFlow API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_HashanEranga_go-task-manager-service_internal_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HashanEranga_go-task-manager-service_internal_models.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate refresh token and logout user
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_HashanEranga_go-task-manager-service_internal_models.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /auth/me:
    get:
      description: Get authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HashanEranga_go-task-manager-service_internal_models.UserResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get new access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_HashanEranga_go-task-manager-service_internal_models.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_HashanEranga_go-task-manager-service_internal_models.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account
      parameters:
      - description: Registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_HashanEranga_go-task-manager-service_internal_models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_HashanEranga_go-task-manager-service_internal_models.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"